+++ cedar-policy/CHANGELOG.md
@@ -423,7 +423,8 @@ Cedar Language Version: 2.0.0
 Cedar Language Version: 2.0.0
 - Initial release of `cedar-policy`.
 
-[Unreleased]: https://github.com/cedar-policy/cedar/compare/v3.0.1...main
+[Unreleased]: https://github.com/cedar-policy/cedar/compare/v3.1.0...main
+[3.1.0]: https://github.com/cedar-policy/cedar/compare/v3.0.1...v3.1.0
 [3.0.1]: https://github.com/cedar-policy/cedar/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/cedar-policy/cedar/compare/v2.4.4...v3.0.0
 [2.4.4]: https://github.com/cedar-policy/cedar/compare/v2.4.3...v2.4.4